ASP-91 First implementation of support for importing channels from a bci2000 file. For now we only do a temporary solution which boils down to importing channel names only.
Wonambi does not support bci2000 fully so we will have to add to this functionality internally to solve the import and reduce dependecy on wonambi.

currently the data inside bci2000 files are very limited as is, which might change if we start using the redwood files instead of bci2000 base files. so for now this is acceptable