ASP-107 First commit for linking the files widget and channel widget together for displaying if a channel is based on a specific recording file.
For this to happen I needed to modify the channel_groups table to now also allow for a field 'based_on_file' which now gets filled in during the creation of the import channel functionality for trc, nev and dat files.

seeing as this is new the marking of file and channel will only happen for entries 2025 or later.

currently if there is a match it will mark them green, still need to add the function for disabling the channel if there is no match between file and channel.