ASP-104 Inclusion of bidso the same way we handled nibabel and wonambi.
This also includes a bit of changes to references as it is no longer an installed dependency.

A small modification to where certain files are stored, and also inclusion of a minimalistic startaspen python file (basically a oneliner that calls main) this will be used for an automatic startup later as a systeminstall.