ASP-123 Creation of the 'Student' rights group which is very similar in function to the 'reader' group with some information masked.
config.py:
Inclusion of the Student group to be looked-up on LDAP

utils.py:
A more general Function for hiding a parameter field for all sessions.

Auto age calculation was not displaying a change, added a little bit of color so the user knows the field has changed based on its input.

interface.py:
- Addition of the Student group ldap
- masking of protocol under files
- masking of DoB in parameters
- hiding of xelo-stem for all sessions