ASP-129 Restructured the layout, so everyone will have the same layout without the dock widgets to the sides and bottom.
everything is now inside the central widget, the docks are removed.

The paramets/event/electrodes/channels/export dock widget has been replaced by the tabview and put permanently on the right.

the bottom filewidget is now also a static place and size.

everything has been resized and the layout has been modified to a QGridLayout.